% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base10to2.R
\name{base10to2}
\alias{base10to2}
\title{Base 10 to 2}
\usage{
base10to2(base10_number)
}
\arguments{
\item{base10_number}{A base 10 natural number}
}
\value{
A vector of integers (1 or 0), with the rightmost element
representing the smallest bit
}
\description{
Converts any natural number (positive integer) to a vector of binary
bits
}
\examples{
base10to2(50)
base10to2(98572)
}
