% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base2to10.R
\name{base2to10}
\alias{base2to10}
\title{Base 2 to 10}
\usage{
base2to10(base2_vector)
}
\arguments{
\item{base2_vector}{A vector of integers (1 or 0), with the rightmost element
representing the smallest bit}
}
\value{
A base 10 natural number
}
\description{
Converts any vector of binary bits to its natural number equivalent
}
\examples{
base2to10(c(0,0,0,0,1,0))
base2to10(c(1,0,0,0,0,0))
base2to10(c(1,0,1,1,0,1))
}
