% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce_date_sequence.R
\name{coerce_date_sequence}
\alias{coerce_date_sequence}
\title{Coerce start/end dates into half-month sequence}
\usage{
coerce_date_sequence(
  start_date,
  end_date,
  time_zone = Sys.timezone(),
  limit = NULL,
  limit_unit = "days",
  ignore.paradox = FALSE
)
}
\arguments{
\item{start_date, end_date}{POSIXt, Date, or a character vector. Attached timezone
doesn't matter, as it is overwritten by the time_zone argument}

\item{time_zone}{a recognized character timezone, optional; defaults to the
local system timezone. The inputs and resulting output will have their
timezone forced to \code{time_zone}'s value}

\item{limit}{a numeric value, optional; determines the furthest in the past
dates will be permitted}

\item{limit_unit}{a character, optional; unit of the limit numeric i.e.
"days", "months". Defaults to "days"}

\item{ignore.paradox}{a boolean which allows the user to create date ranges
which go into the future, so as to e.g. proactively create data folders}
}
\value{
A vector of Date values which includes the \code{start_date}, \code{end_date},
and any 1st and/or 16th of the months in-between
}
\description{
The coerce_date_sequence function accepts start and end dates in the format
year-mo-dy, pushing the start date into the past to the nearest valid 1st or
16th of the month (inclusive) and pushes the end date into the past to the
nearest valid 1st of 16th of the month (exclusive). It then breaks the period
into two-week chunks that can be looped through to download multiple two-week
datasets.
}
\examples{
start_date <- as.Date("2023-01-01")
end_date <- as.Date("2024-01-01")
download_dates <- coerce_date_sequence(
  start_date = start_date,
  end_date = end_date,
  time_zone = "America/Chicago"
)
download_dates
}
