% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_padata.R
\name{get_padata}
\alias{get_padata}
\title{GET latest PurpleAir sensor data}
\usage{
get_padata(sensors, fields, api_read_key = get_api_key())
}
\arguments{
\item{sensors}{Either a \code{vector} or \code{dataframe} specifying sensors to GET
\itemize{
\item vector: a \code{vector} of \code{numeric} sensor indexes
\item dataframe: if any sensors are private, read keys will be needed to GET
their data. In this case, a dataframe with \code{numeric} sensor_index and
\code{character} read_key columns will be needed
}}

\item{fields}{A \verb{character vector} of api fields to return; valid fields can
be found at https://api.purpleair.com/#api-sensors-get-sensor-data and
https://community.purpleair.com/t/api-fields-descriptions/4652}

\item{api_read_key}{A valid \code{character} PurpleAir api read key. The function
will grab an existing read key if saved to your R environment with
\code{set_api_key()}}
}
\value{
A dataframe with field values for each supplied sensor
}
\description{
\code{get_padata()} allows a user to easily GET the latest field data for a
group of sensors by supplying sensor identifying information. Uses the
PurpleAir Get Sensor Data api request to pull data from PurpleAir
}
\examples{
# Dummy example inputs which will fail if actually run
\dontrun{
sensors <- data.frame(
  sensor_index = 122948,
  read_key = "28qhig4ghqh290qt"
)
fields_vector <- c("name", "uptime")
api_read_key <- "EHAIGEH1-18Y9-81H8-GGI9-HTQ238HQ9H8H"
get_padata(sensors, fields_vector, api_read_key)
}
}
