% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not_in.R
\name{\%notin\%}
\alias{\%notin\%}
\title{Find if an element in Not In something}
\usage{
x \%notin\% table
}
\arguments{
\item{x}{vector or NULL: the values to be matched. Long vectors are
supported.}

\item{table}{vector or NULL: the values to be matched against. Long vectors
are not supported.}
}
\value{
A vector same length as x
}
\description{
Performs the reverse of the \%in\% operator, which is a binary operator of the
\code{\link[base:match]{base::match()}} function, returning \code{TRUE} if a match is not found
}
\examples{
1:10 \%notin\% c(1, 3, 5, 9)
}
\seealso{
\code{\link[=match]{match()}} which this function is the reverse of
}
