% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_padata.R
\name{load_padata}
\alias{load_padata}
\title{Load saved PurpleAir data}
\usage{
load_padata(
  data_dir,
  load_interval,
  indexes = NULL,
  names = NULL,
  projects = NULL,
  all_sensors = FALSE,
  level = c("0", "1", "2"),
  letter_mod = "default",
  data_suffix = "default",
  verbose = FALSE
)
}
\arguments{
\item{data_dir}{A \code{character} path from pc root to directory containing pa
data folders and \code{PA_Metadata.csv} sheet. The values in the metadata
\code{sensor_index}, \code{Name}, and \code{Project} columns are referenced in this
function's matching arguments}

\item{load_interval}{A lubridate date \code{\link[lubridate:interval]{lubridate::interval()}} that loads any
data files that fall within it}

\item{indexes}{A \code{vector} of \code{character} or \code{numeric} PA data indexes. Valid
indexes are always six digits long, but may include a seventh digit some time
in the future}

\item{names}{A \code{vector} of \code{character} PA sensor names. This argument and
\code{projects} accept partial, case-insensitive matching}

\item{projects}{A \code{vector} of \code{character} PA projects. This argument and
\code{names} accept partial, case-insensitive matching}

\item{all_sensors}{Default: \code{FALSE}. An optional \code{boolean} which confirms no
filtering. If data from all sensors is desired, input no values for
indexes/names/projects and set this to \code{TRUE}}

\item{level}{Default: \code{2}. An optional \code{integer} or \code{character} which
specifies the level of data to be loaded. 0: Raw data. 1: Quality-assured
data. 2: Analysis-ready data}

\item{letter_mod}{An optional \code{character} that will be appended to the level
to specify an alternate directory. e.g. "B" -> "2B_ANALYSIS"}

\item{data_suffix}{An optional \code{character} that will be used instead of the
default for a certain level file suffix. Defaults for 0, 1, and 2 levels are
"RAW", "VAL", and "1HR", respectively}

\item{verbose}{Default: \code{FALSE}. An optional \code{boolean} which disables or
enables cli progress bar updating on loading progress. Function will still
give parameter feedback if incorrect variables are supplied}
}
\value{
A \code{dataframe} containing all pa data matching selection criteria
}
\description{
Loads PurpleAir(PA) sensor data based on the union of input sensor
indexes, names, and projects. Stitches files into dataframe for analysis or
data sharing.
}
\examples{
\dontrun{
load_padata()
}
}
