% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_path.R
\name{optimize_path}
\alias{optimize_path}
\title{Optimize file path}
\usage{
optimize_path(path, feedback = FALSE)
}
\arguments{
\item{path}{a character file path constructed of folder names separated by
"/" separators}

\item{feedback}{An optional boolean which suppresses/allows feedback on
removed path elements}
}
\value{
A character file path that describes shortest possible path to a
directory
}
\description{
\code{optimize_path()} detects and removes elements from a file path using ".."
characters to perform up-directory moves, assuming a "/x/y/z" file path type.
Can be useful for shortening file paths created by \code{here()} function.
}
\examples{
path <- paste(
  "C:", "Users", "CoolUser",
  "ReallyLongFolderWeDontUse", "ReallyLongFolderWeDontUse2ElectricBoogaloo",
  "..", "..", "FolderA",
  sep = "/"
  )
short_path <- optimize_path(path)
short_path
}
