% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_excel_allsheets.R
\name{read_excel_allsheets}
\alias{read_excel_allsheets}
\title{Read all sheets in xlsx to list}
\usage{
read_excel_allsheets(filename)
}
\arguments{
\item{filename}{A character filepath ending with the excel file to be loaded}
}
\value{
A list of dataframes, one per sheet
}
\description{
Functions similarly to \code{\link[readxl:readxl-package]{readxl::readxl()}} in that it bundles the function to apply
on every sheet of an excel file, loading the results into a list with a
member per sheet
}
\examples{
\dontrun{
read_excel_allsheets("C:/MyExcel.xlsx")
}
}
