% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saair-package.R
\docType{package}
\name{saair-package}
\alias{saair}
\alias{saair-package}
\title{saair: Useful code for Dr. Brown Air Quality lab at UTSA}
\description{
Helps with tasks common to coding for Dr. Brown & Dr. Lopez-Ochoa heat and air quality research and community partnership activities. Includes several functions sourced from Stack Overflow as well as many longer original works.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tarias7274/saair}
  \item Report bugs at \url{https://github.com/tarias7274/saair/issues}
}

}
\author{
\strong{Maintainer}: Tristan Arias \email{tristan.arias@utsa.edu} (\href{https://orcid.org/0009-0002-3329-9216}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Kristen Brown \email{kristen.brown2@utsa.edu} (\href{https://orcid.org/0000-0002-0201-0558}{ORCID}) [copyright holder]
}

}
\keyword{internal}
