% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_api_key.R
\name{set_api_key}
\alias{set_api_key}
\title{Set api key}
\usage{
set_api_key(key = NULL)
}
\arguments{
\item{key}{Kept NULL in order to use the obfuscating askpass interface}
}
\value{
An action setting the env var \code{SAAIR_KEY} to input key value
}
\description{
Prompts user to set a PurpleAir API key. Typically follows this example format:
B144C732-11DE-11EE-A445-42098A784009
}
\examples{
set_api_key()
}
