% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calpat.R
\name{calpat}
\alias{calpat}
\title{Calibrate PurpleAir temperatures}
\usage{
calpat(
  data,
  models = models,
  calibration = c("sunlit", "shaded", "indoor", "shell"),
  temperature_col = "T_air",
  rh_col = "RH",
  time_col = "Timestamp_Local",
  output_c_col = "calT_C",
  output_f_col = "calT_F",
  timezone = "America/Chicago",
  timestamp_is_char = TRUE,
  warn_range = TRUE
)
}
\arguments{
\item{data}{A dataframe of PurpleAir data, such as output from
\code{load_padata()}}

\item{models}{Optional list of calibration models. If none are supplied, the
package's internal calibrations are used}

\item{calibration}{A character that best describes the deployment conditions.
All sensors in \code{data} dataframe should have the same condition
"sunlit": sensor is deployed outdoors and is largely sunlit during the day
"shaded": sensor is deployed outdoors and is largely shaded during the day
"indoor": sensor is deployed indoors
"shell" (unimplemented): sensor is contained in a 3D-printed shell
deployed at a Via bus stop}

\item{temperature_col}{A character specifying the name of the Fahrenheit data
column if it's not "T_air"}

\item{rh_col}{A character specifying the name of the 0-100\% relative humidity
data if it's not "RH"}

\item{time_col}{A character specifying the name of the timestamp column if
it's not "Timestamp_Local". See also timestamp_is_char}

\item{output_c_col}{A character specifying the name of the output Celsius
column. Default: "calT_C"}

\item{output_f_col}{A character specifying the name of the output Fahrenheit
column. Default: "calT_F"}

\item{timezone}{The local time zone for the data}

\item{timestamp_is_char}{If \code{TRUE}, \code{time_col} is treated as a local
clock time even if R has assigned the wrong timezone attribute. This
matches current \code{load_padata()} behavior.}

\item{warn_range}{Get warnings if it looks like your temperature or RH units
are wrong}
}
\value{
A dataframe with calT_C and calT_F (calibrated temperatures
in C and F), PA_time_local, and PA_time_UTC appended to original data
}
\description{
Calibrate PurpleAir temperatures
}
\examples{
\dontrun{
start_date <- as_datetime("2025-04-20", tz = "America/Chicago")
end_date <- as_datetime("2025-04-21", tz = "America/Chicago")

pa_data <- load_padata(
  data_dir = PAfiles,
  load_interval = interval(start = start_date, end = end_date),
  indexes = c(150840, 151664)
  )

models <- readRDS("PA_temperature_calibration_models.rds")

pa_calibrated <- calpat(
  data = pa_data,
  models = models,
  calibration = "shaded"
  )

ggplot(pa_calibrated) +
  theme_bw() +
  geom_point(aes(x = PA_time_local, y = T_air), color = "purple") +
  geom_point(aes(x = PA_time_local, y = calT_F), color = "forestgreen") +
  geom_point(aes(x = PA_time_local, y = calT_C), color = "blue") +
  facet_wrap(~sensor_index) +
  xlim(start_date, end_date)
}
}
