% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_api_key.R
\name{get_api_key}
\alias{get_api_key}
\title{Get api key}
\usage{
get_api_key()
}
\value{
Value for SAAIR_KEY environment variable if it exists in the user's environment
}
\description{
Get api key
}
\examples{
get_api_key()
}
