% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sensor_data.R
\name{get_sensor_data}
\alias{get_sensor_data}
\title{GET PurpleAir Sensor Data}
\usage{
get_sensor_data(sensors_df, fields, api_read_key)
}
\arguments{
\item{sensors_df}{A dataframe with PurpleAir sensor info}

\item{fields}{A character vector of api fields to return}

\item{api_read_key}{A valid character PurpleAir api read key
inheritParams}
}
\value{
A dataframe with field values
}
\description{
GET PurpleAir Sensor Data
}
\examples{
sensors <- data.frame(
  sensor_index = 122948,
  MAC_SN = "00:00:00:00:00:00",
  Name = "PurpleSensor-0089",
  read_key = "28qhig4ghqh290qt"
)
fields_vector <- c("name", "uptime")
api_read <- "EHAIGEH1-18Y9-81H8-GGI9-HTQ238HQ9H8H"
get_sensor_data(sensors, fields_vector, api_read)
}
