% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_padata.R
\name{load_padata}
\alias{load_padata}
\title{Load saved PurpleAir data}
\usage{
load_padata(
  data_dir,
  load_interval,
  indexes = NULL,
  pa_names = NULL,
  projects = NULL,
  all_sensors = FALSE
)
}
\arguments{
\item{data_dir}{A character path from pc root to directory containing pa data
folders}

\item{load_interval}{A lubridate interval that loads any data files which
have intervals which overlap with it}

\item{indexes}{A vector of character or numeric pa data indexes}

\item{pa_names}{A vector of character pa names}

\item{projects}{A vector of character pa projects}

\item{all_sensors}{A boolean which confirms no filtering. If data from all
sensors is desired, input no values for indexes/pa_names/projects and set
this to \code{TRUE}}
}
\value{
A dataframe containing all pa data matching selection criteria
}
\description{
Loads raw purpleair(PA) sensor data based on the union of input sensor
indexes, names, and project. Stitches files into dataframe for analysis or
output.
}
\examples{
\dontrun{
load_padata()
}
}
